+++ src/index.ts
@@ -6,23 +6,6 @@ import { HTTPException } from "hono/http-exception"
 import { constructResponse } from "./utils/response"
 import { trimTrailingSlash } from "hono/trailing-slash"
 
-export type Bindings = {
-  DATABASE_URL: string
-  COMULINE_ENV: string
-  UPSTASH_REDIS_REST_TOKEN: string
-  UPSTASH_REDIS_REST_URL: string
-}
-
-export type Variables = {
-  db: Database<Bindings>["db"]
-  constructResponse: typeof constructResponse
-}
-
-export type Environments = {
-  Bindings: Bindings
-  Variables: Variables
-}
-
 const api = createAPI()
 
 const app = api
