+++ src/modules/api.ts
@@ -1,5 +1,5 @@
 import { OpenAPIHono } from "@hono/zod-openapi"
-import { Environments } from ".."
+import { type Environments } from "@/type"
 
 export const createAPI = <T extends Environments>() =>
   new OpenAPIHono<T>({ strict: true })
